feat(pvm2): delete remaining legacy PVM transpiler + drop bitmask field
Phase 3b: PVM2 cleanup follow-on to commit e56868ab.

* Delete remaining PVM-only modules in javm-transpiler: `assembler.rs`,
  `emitter.rs`, `linker.rs`, `program.rs`, `riscv.rs`, plus the
  peephole passes in `lib.rs`. Extract the shared ELF parsing helpers
  (`LinkedElf`, `parse_linked_elf`, `find_all_section_bytes`) into a
  new `elf.rs` module used by `linker_rv.rs`.
* Drop `Image.packed_bitmask` and `ImageCap.bitmask` fields. Both
  were exclusively populated by the legacy linker and not consumed
  by any PVM2 code path.
* Drop `Compiler::new`'s now-unused `bitmask: &[u8]` and `jump_table:
  &[u32]` parameters.
* Rename `build_javm::build_pvm2` → `build` (the legacy `build` and
  `build_service` were dead since Phase 2). Delete the legacy
  `riscv64em-javm.json` target spec. Rename the bench `_PVM2_BLOB`
  env vars → `_BLOB` and the on-disk extension `.pvm2` → `.pvm`.
* Drop the `ERR_ENDPOINT_UNDEFINED` rejection in `nub-arch-x86`'s
  call loop — `entry_pc == 0` is a legitimate PC under PVM2.
* Migrate the few remaining hand-authored PVM bytecodes in tests to
  PVM2 `ecalli 0` (0x0000_200B) / `jal x0, 0` (0x0000_006F).
* Tidy a handful of pre-existing clippy warnings that became hard
  errors once strict mode caught them.

`cargo build --workspace`, `cargo test --workspace -- --skip workloads`,
`cargo clippy --workspace --all-targets -- -D warnings`,
`cargo run --release -p javm-bench --example pvm2_smoke` (12/12 PASS
with bit-identical gas between interp and recomp).
