--- rust/build-javm/src/riscv64em-javm.json
@@ -1,27 +1,0 @@
-{
-  "arch": "riscv64",
-  "cpu": "generic-rv64",
-  "crt-objects-fallback": "false",
-  "data-layout": "e-m:e-p:64:64-i64:64-i128:128-n32:64-S64",
-  "eh-frame-header": false,
-  "emit-debug-gdb-scripts": false,
-  "features": "+e,+m,+zbb,+xtheadcondmov",
-  "linker": "rust-lld",
-  "linker-flavor": "ld.lld",
-  "llvm-abiname": "lp64e",
-  "llvm-target": "riscv64",
-  "max-atomic-width": 0,
-  "panic-strategy": "abort",
-  "relocation-model": "pie",
-  "target-pointer-width": 64,
-  "singlethread": true,
-  "exe-suffix": ".elf",
-  "os": "none",
-  "env": "javm",
-  "pre-link-args": {
-    "ld": [
-      "--emit-relocs",
-      "--unique"
-    ]
-  }
-}
