+++ rust/javm-bench/examples/pvm2_disasm.rs
@@ -12,13 +12,13 @@ fn main() {
         .nth(1)
         .unwrap_or_else(|| "goldilocks_mul".into());
     let blob_bytes: &[u8] = match which.as_str() {
-        "goldilocks_mul" => include_bytes!(env!("GOLDILOCKS_MUL_PVM2_BLOB")),
-        "poly_eval" => include_bytes!(env!("POLY_EVAL_PVM2_BLOB")),
-        "ed25519" => include_bytes!(env!("ED25519_PVM2_BLOB")),
-        "poseidon2_perm" => include_bytes!(env!("POSEIDON2_PERM_PVM2_BLOB")),
-        "mini_verifier" => include_bytes!(env!("MINI_VERIFIER_PVM2_BLOB")),
-        "fri_fold_tree" => include_bytes!(env!("FRI_FOLD_TREE_PVM2_BLOB")),
-        "ecrecover" => include_bytes!(env!("ECRECOVER_PVM2_BLOB")),
+        "goldilocks_mul" => include_bytes!(env!("GOLDILOCKS_MUL_BLOB")),
+        "poly_eval" => include_bytes!(env!("POLY_EVAL_BLOB")),
+        "ed25519" => include_bytes!(env!("ED25519_BLOB")),
+        "poseidon2_perm" => include_bytes!(env!("POSEIDON2_PERM_BLOB")),
+        "mini_verifier" => include_bytes!(env!("MINI_VERIFIER_BLOB")),
+        "fri_fold_tree" => include_bytes!(env!("FRI_FOLD_TREE_BLOB")),
+        "ecrecover" => include_bytes!(env!("ECRECOVER_BLOB")),
         _ => panic!("unknown guest"),
     };
     let img = Image::from_ssz_bytes(blob_bytes).unwrap();
