+++ rust/javm-bench/tests/workloads.rs
@@ -49,7 +49,7 @@ fn check_workload(name: &str, blob: &[u8], expected_value: u64, expected_gas: u6
 fn prime_sieve() {
     check_workload(
         "prime_sieve",
-        include_bytes!(env!("PRIME_SIEVE_PVM2_BLOB")),
+        include_bytes!(env!("PRIME_SIEVE_BLOB")),
         0x2578,
         8_966_290,
     );
@@ -59,7 +59,7 @@ fn prime_sieve() {
 fn ed25519() {
     check_workload(
         "ed25519",
-        include_bytes!(env!("ED25519_PVM2_BLOB")),
+        include_bytes!(env!("ED25519_BLOB")),
         0x1,
         2_362_040,
     );
@@ -69,7 +69,7 @@ fn ed25519() {
 fn keccak() {
     check_workload(
         "keccak",
-        include_bytes!(env!("KECCAK_PVM2_BLOB")),
+        include_bytes!(env!("KECCAK_BLOB")),
         0x39e5_0259,
         101_642,
     );
@@ -79,7 +79,7 @@ fn keccak() {
 fn blake2b() {
     check_workload(
         "blake2b",
-        include_bytes!(env!("BLAKE2B_PVM2_BLOB")),
+        include_bytes!(env!("BLAKE2B_BLOB")),
         0xee1f_55f1,
         63_192,
     );
@@ -89,7 +89,7 @@ fn blake2b() {
 fn ecrecover() {
     check_workload(
         "ecrecover",
-        include_bytes!(env!("ECRECOVER_PVM2_BLOB")),
+        include_bytes!(env!("ECRECOVER_BLOB")),
         0x1,
         6_819_891,
     );
@@ -99,7 +99,7 @@ fn ecrecover() {
 fn goldilocks_mul() {
     check_workload(
         "goldilocks_mul",
-        include_bytes!(env!("GOLDILOCKS_MUL_PVM2_BLOB")),
+        include_bytes!(env!("GOLDILOCKS_MUL_BLOB")),
         0x2cf7_3e57,
         2_400_154,
     );
@@ -109,7 +109,7 @@ fn goldilocks_mul() {
 fn poseidon2_perm() {
     check_workload(
         "poseidon2_perm",
-        include_bytes!(env!("POSEIDON2_PERM_PVM2_BLOB")),
+        include_bytes!(env!("POSEIDON2_PERM_BLOB")),
         0x3ce3_3156,
         14_561_189,
     );
@@ -119,7 +119,7 @@ fn poseidon2_perm() {
 fn mini_verifier() {
     check_workload(
         "mini_verifier",
-        include_bytes!(env!("MINI_VERIFIER_PVM2_BLOB")),
+        include_bytes!(env!("MINI_VERIFIER_BLOB")),
         0xf98f_c4ab,
         5_878_907,
     );
@@ -129,7 +129,7 @@ fn mini_verifier() {
 fn poly_eval() {
     check_workload(
         "poly_eval",
-        include_bytes!(env!("POLY_EVAL_PVM2_BLOB")),
+        include_bytes!(env!("POLY_EVAL_BLOB")),
         0x01da_34e2,
         9_003_090,
     );
@@ -139,7 +139,7 @@ fn poly_eval() {
 fn fri_fold_tree() {
     check_workload(
         "fri_fold_tree",
-        include_bytes!(env!("FRI_FOLD_TREE_PVM2_BLOB")),
+        include_bytes!(env!("FRI_FOLD_TREE_BLOB")),
         0x37e6_76f4,
         6_189_475,
     );
