+++ rust/javm-cap/tests/cap_smoke.rs
@@ -13,7 +13,6 @@ fn make_image_cap() -> ImageCap {
     let code: Vec<u8> = vec![0xAB, 0xCD];
     ImageCap {
         code,
-        bitmask: Vec::new(),
         jump_table: Vec::new(),
         jump_table_offsets: Vec::new(),
         endpoints: Vec::new(),
