+++ rust/javm-cap/tests/image.rs
@@ -20,8 +20,8 @@ fn empty_image_hashes_deterministically() {
 fn image_ssz_roundtrip() {
     let mut img = Image::empty();
     img.code = b"sample code".to_vec();
-    img.packed_bitmask = vec![0xFF, 0x07]; // 11 bits set, all-1s
     img.jump_table = vec![0u32, 4, 8];
+    img.jump_table_offsets = vec![0, 3];
     img.endpoints.insert(
         0,
         EndpointDef {
