+++ rust/javm/src/ecall.rs
@@ -1233,8 +1233,9 @@ mod tests {
         let mut vm = fixture_vm();
         let mut cache = CacheDirectory::new();
         let mut img = Image::empty();
-        img.code = vec![10u8, 0];
-        img.packed_bitmask = vec![0b01u8];
+        // PVM2 `ecalli 0` — 32-bit custom-0 word at PC 0.
+        img.code = 0x0000_200Bu32.to_le_bytes().to_vec();
+        img.jump_table_offsets = vec![0, 0];
         let image_hash = cache
             .put_cap(&Cap::image_with_slots(&img, &[], &[]).unwrap())
             .unwrap();
@@ -1254,7 +1255,7 @@ mod tests {
         );
         let running = vm.stack.running_instance().unwrap();
         assert_eq!(running.image_hash, image_hash);
-        assert_eq!(running.program.code, vec![10u8, 0]);
+        assert_eq!(running.program.code, 0x0000_200Bu32.to_le_bytes().to_vec());
     }
 
     #[test]
@@ -1557,8 +1558,9 @@ mod tests {
 
         // Publish a tiny child image with no pinned/initial slots.
         let mut child_img = javm_cap::image::Image::empty();
-        child_img.code = vec![10u8, 0]; // ecalli 0
-        child_img.packed_bitmask = vec![0b01u8];
+        // PVM2 `ecalli 0` — 32-bit custom-0 word at PC 0.
+        child_img.code = 0x0000_200Bu32.to_le_bytes().to_vec();
+        child_img.jump_table_offsets = vec![0, 0];
         let image_hash = cache
             .put_cap(&Cap::image_with_slots(&child_img, &[], &[]).unwrap())
             .unwrap();
@@ -1647,8 +1649,8 @@ mod tests {
         // Publish a no-op image (one Halt instruction) + empty cnode
         // + Cap::Instance referencing them.
         let mut child_img = javm_cap::image::Image::empty();
-        child_img.code = vec![10u8, 0];
-        child_img.packed_bitmask = vec![0b01u8];
+        child_img.code = 0x0000_200Bu32.to_le_bytes().to_vec();
+        child_img.jump_table_offsets = vec![0, 0];
         let image_hash = cache
             .put_cap(&Cap::image_with_slots(&child_img, &[], &[]).unwrap())
             .unwrap();
